Remove stale performance check
Now the performance check becomes stale. It always fails now. It looks
like the reason is we didn't update something in time. But I feel it is
not good to monitoring the performance in github CI since it is too
unstable, which makes it menaingless.

So I decide to remove the performance monitoring now.