Reduce the size of benchmarks
Previously it took too long to run the benchmarks. But the longer
benchmarking time didn't take any new insight for us. So I think it is
not worthy to give such a long time to do the benchmarking. This patch
mitigates the problem by reduing the size of of the benchmarks.