Define ASYNC_SIMPLE_HAS_NOT_AIO in source if we can't find libaio.h
While we handled ASYNC_SIMPLE_HAS_NOT_AIO carefully in the build
systems, it is still not so friendly if we treat async_simple as a
header only library. So this patch tries to define
ASYNC_SIMPLE_HAS_NOT_AIO in the source directly if it can't find the libaio.h.