Hacking in the clang/cmake CI to downgrade consteval to constexpr
Currently, there is a problem in either clang14 or libstdc++13 in the
CI environment that complaining that there is a function marked as
consteval but not constant expression.

We add a `-Dconsteval=constexpr` in the CI hackly to make that CI green.
We need to remove it when the environment changes someday.